FEBio2: Moved Timer and PlotFile class to FEBioLib
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBio2@2876 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBio2/FETriphasic.cpp
@@ -125,7 +125,7 @@ double FETriphasic::Concentration(FEMaterialPoint& pt, const int ion)
 	// solubility
 	double kappa = m_pSolute[ion]->m_pSolub->Solubility(pt);
 	// charge number
-	int z = m_pSolute[ion]->ChargeNumber();
+	int z = (int) m_pSolute[ion]->ChargeNumber();
 	
 	// actual concentration
 	double ca = kappa*c*exp(-z*m_Fc*psi/m_Rgas/m_Tabs);
